@@ -1,8 +1,9 @@
 module github.com/go-delve/delve
 
-go 1.11
+go 1.16
 
 require (
+	github.com/aquasecurity/libbpfgo v0.1.2-0.20210708203834-4928d36fafac
 	github.com/cosiner/argv v0.1.0
 	github.com/creack/pty v1.1.9
 	github.com/google/go-dap v0.5.0
@@ -14,7 +15,7 @@ require (
 	github.com/spf13/cobra v1.1.3
 	go.starlark.net v0.0.0-20200821142938-949cc6f4b097
 	golang.org/x/arch v0.0.0-20190927153633-4e8777c89be4
-	golang.org/x/sys v0.0.0-20200625212154-ddb9806d33ae
+	golang.org/x/sys v0.0.0-20210514084401-e8d321eab015
 	golang.org/x/tools v0.0.0-20191127201027-ecd32218bd7f
 	gopkg.in/yaml.v2 v2.4.0
 )
