@@ -1,6 +1,9 @@
 package proc
 
-import "github.com/go-delve/delve/pkg/elfwriter"
+import (
+	"github.com/go-delve/delve/pkg/elfwriter"
+	"github.com/go-delve/delve/pkg/proc/internal/ebpf"
+)
 
 // Process represents the target of the debugger. This
 // target could be a system process, core file, etc.
@@ -42,11 +45,16 @@ type ProcessInternal interface {
 	WriteBreakpoint(*Breakpoint) error
 	EraseBreakpoint(*Breakpoint) error
 
+	SupportsBPF() bool
+	SetUProbe(string, []ebpf.UProbeArgMap) error
+
 	// DumpProcessNotes returns ELF core notes describing the process and its threads.
 	// Implementing this method is optional.
 	DumpProcessNotes(notes []elfwriter.Note, threadDone func()) (bool, []elfwriter.Note, error)
 	// MemoryMap returns the memory map of the target process. This method must be implemented if CanDump is true.
 	MemoryMap() ([]MemoryMapEntry, error)
+
+	GetBufferedTracepoints() []ebpf.RawUProbeParams
 }
 
 // RecordingManipulation is an interface for manipulating process recordings.
